feat(frontend): separates the code in different components
